[arccore,alina] Cleanup file 'AlinaUtils.cc' and move class 'PropertyTreeImpl' in this file.
